[REFACTOR] mainXXX 를 home으로 변경했어요.
